Added restrictions to edit and delete

+++ RealTime/food/views.py
@@ -48,13 +48,16 @@ def update_item(request,id):
 
         return render(request,'food/item_form.html',{'form':form,'item':item})
     else:
-        return redirect('food:index')
+        return HttpResponse('<h1>You have no permission to edit this item</h1>')
 
 def delete_item(request,id):
     item = Item.objects.get(id=id)
     
-    if request.method == 'POST':
-        item.delete()
-        return redirect('food:index')
+    if item.user_name.get_username == request.user.get_username:
+        if request.method == 'POST':
+            item.delete()
+            return redirect('food:index')
         
-    return render(request,'food/item_delete.html',{'item':item})
+        return render(request,'food/item_delete.html',{'item':item})
+    else:
+        return HttpResponse('<h1>You have no permission to delete this item</h1>')
